refactor(soundpacks): split community pack model
Move CommunityPack metadata into a focused service model module while preserving the original service import path.

Tested: uv run pytest -q -n 0 --tb=short tests/test_soundpack_paths.py tests/test_settings_dialog_audio_events.py; uv run ruff check src tests scripts; uv run pyright

Confidence: high
